+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## Cldr Locale Display v0.1.0
+## Cldr Locale Display v1.0.0
 
-This is the changelog for Cldr Locale Display v0.1.0 released on ______, 2021.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_locale_display/tags)
+This is the changelog for Cldr Locale Display v1.0.0 released on July 1st, 2021.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_locale_display/tags)
+
+### Enhancements
+
+This is the first release of `ex_cldr_locale_display` which implements the [CLDR locale display name algorithm](https://unicode-org.github.io/cldr/ldml/tr35-general.html#locale_display_name_algorithm).
+